수정: 신청한 운동 프로그램 목록 추가:
