+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,6 @@
         <activity android:name=".TrainerExrProgramActivity" />
         <activity android:name=".MemberAllExrProgramListActivity" />
         <activity android:name=".ExrProgramDetailActivity" />
-        <activity android:name=".MemberExrProgramListActivity" />
         <activity android:name=".AdminVideoManageActivity" />
         <activity android:name=".TrainerVideoListActivity" />
         <activity android:name=".TrainerVideoRegActivity" />
